+++ packages/dart_neo4j/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add optional `timeout` parameter to `BoltConnection.run` (and `BoltConnection._sendMessage`, with default set to 30 seconds for backward compatibility); the timeout is used on the `pull` message.
 - Add optional `timeout` parameter to `SessionConfiguration` (defaults to 30 seconds) and use that timeout when running queries directly over the session.
 - Use `TransactionConfig.timeout` when running queries over a transaction.
+- Implement `startQuery()` methods in addition to the existing `run()` methods. The existing implementations of `run()` returns only after all records have been streamed from the server. `startQuery()` returns before the query has fully executing, allowing for early streaming of records by clients. The query has finished executing when [Result.done] and [Result.summary()] complete (they are based on the same future).
 
 ## 1.2.0